Restructured the packages
