+++ First-Java-Prog/src/com/condition/loop/basicQ/Calculator.java
@@ -1,4 +1,4 @@
-package com.condition.loop;
+package com.condition.loop.basicQ;
 
 import java.util.Scanner;
